Refactor - adding common resolver
