--- server/lib/resolve/title.js
@@ -1,3 +1,0 @@
-'use strict';
-
-module.exports = exports = item => item;
